Moved CompleteYourProfile related files to own folder.
