+++ src/js/components/CompleteYourProfile/PersonalizedScoreIntroModal.jsx
@@ -9,7 +9,7 @@ import IconButton from '@material-ui/core/IconButton';
 import { withStyles, withTheme } from '@material-ui/core/styles';
 import { hasIPhoneNotch } from '../../utils/cordovaUtils';
 import { renderLog } from '../../utils/logging';
-import CandidateItem from './CandidateItem';
+import CandidateItem from '../Ballot/CandidateItem';
 import VoterActions from '../../actions/VoterActions';
 import VoterConstants from '../../constants/VoterConstants';
