+++ src/js/components/Navigation/HeaderBar.jsx
@@ -13,23 +13,23 @@ import AccountCircleIcon from '@material-ui/icons/AccountCircle';
 import { withStyles } from '@material-ui/core/styles';
 import Badge from '@material-ui/core/Badge';
 import { historyPush, isWebApp, isCordova, hasIPhoneNotch } from '../../utils/cordovaUtils';
-import AdviserIntroModal from '../Ballot/AdviserIntroModal';
+import AdviserIntroModal from '../CompleteYourProfile/AdviserIntroModal';
 import AppActions from '../../actions/AppActions';
 import AppStore from '../../stores/AppStore';
 import BallotActions from '../../actions/BallotActions';
 import cookies from '../../utils/cookies';
-import FirstPositionIntroModal from '../Ballot/FirstPositionIntroModal';
+import FirstPositionIntroModal from '../CompleteYourProfile/FirstPositionIntroModal';
 import FriendStore from '../../stores/FriendStore';
 import HeaderBarProfilePopUp from './HeaderBarProfilePopUp';
 import HeaderBarLogo from './HeaderBarLogo';
-import HowItWorksModal from '../Ballot/HowItWorksModal';
+import HowItWorksModal from '../CompleteYourProfile/HowItWorksModal';
 import { renderLog } from '../../utils/logging';
 import OrganizationActions from '../../actions/OrganizationActions';
 import PaidAccountUpgradeModal from '../Settings/PaidAccountUpgradeModal';
-import PersonalizedScoreIntroModal from '../Ballot/PersonalizedScoreIntroModal';
+import PersonalizedScoreIntroModal from '../CompleteYourProfile/PersonalizedScoreIntroModal';
 import SelectBallotModal from '../Ballot/SelectBallotModal';
 import SignInModal from '../Widgets/SignInModal';
-import ValuesIntroModal from '../Ballot/ValuesIntroModal';
+import ValuesIntroModal from '../CompleteYourProfile/ValuesIntroModal';
 import VoterGuideActions from '../../actions/VoterGuideActions';
 import VoterSessionActions from '../../actions/VoterSessionActions';
 import VoterStore from '../../stores/VoterStore';
